Fix person class
